requirements.txt
# Jonty - Personal AI Assistant
# Phase 1: Data Engine Dependencies

# Core ML/AI
faiss-cpu>=1.13.0
sentence-transformers>=2.6.0
numpy>=1.26.0

# File Processing
PyMuPDF>=1.23.0
python-docx>=0.8.10
pandas>=2.0.0
openpyxl>=3.1.0

# Configuration
pyyaml>=6.0

# Optional: For audio/speech (Phase 5+)
# openai-whisper==20230314

# Optional: For LLM (Phase 2+)
# ollama==0.0.1
# transformers==4.30.0

# Optional: OCR support (Phase 1+)
# pytesseract==0.3.10

# Development
python-dotenv==1.0.0
